Changes to pass units on the url to highcharts.php
These changes, no longer require highcharts.php  to load chartslivedata.php so this page has no php on it. It can be used as a regular html page to display any support highchart chart. Changes to index.php were required to pass the units on the url for highcharts.php page.